Implement division using repeated subtraction
I took two variables. In the while loop, I added a condition that checks if no2 (the bigger number) is greater than or equal to no1. Inside the loop, no2 is reduced by no1. When no2 becomes smaller than no1, the loop stops. Each time the loop runs, the result is increased by 1.